setup modules are renamed to config
